requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
Babel==2.9.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3.1
cryptography==3.3.1
daphne==3.0.1
distlib==0.3.1
Django==3.1.5
django-admin-rangefilter==0.6.4
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-filters==1.0.0.dev2
docutils==0.16
drf-yasg==1.20.0
filelock==3.0.12
flake8==3.8.4
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
imagesize==1.2.0
incremental==17.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.8
pluggy==0.13.1
priority==1.3.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.4
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2020.5
requests==2.25.1
ruamel.yaml==0.16.12
service-identity==18.1.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
toml==0.10.2
tox==3.21.2
Twisted==20.3.0
txaio==20.12.1
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.4.0
zope.interface==5.2.0
